requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2

sqlmodel==0.0.24
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10

python-jose[cryptography]==3.4.0

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic==2.11.4
pydantic-settings==2.9.1
email-validator==2.2.0

redis==6.0.0
celery==5.4.0

python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2

pytest==8.3.5
httpx==0.28.1

alembic==1.15.2